data live streaming is working (api is a bit delayed) and test client included

+++ src/logic/HandshakeHello.cpp
@@ -4,7 +4,7 @@
 #include <openssl/rand.h>
 
 
-Gut::HandShakeHello::HandShakeHello(std::shared_ptr<Client> &client, uint64_t reqId  , String clientpublickey)
+Gut::HandShakeHello::HandShakeHello(std::shared_ptr<Client> &client, uint32_t reqId  , String clientpublickey)
 	: Task(client, reqId), clientPublicKey(std::move(clientpublickey))
 {
 	std::cout << "handshakehello started" << std::endl;
